requirements.txt


Django==3.0.7
django-allauth==0.41.0
django-cors-headers==3.2.1
django-node==4.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-storages==1.9.1
django-taggit==1.2.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0

virtualenv==16.4.0

